Move relay to separate impl
