+++ ntied-transport/src/lib.rs
@@ -3,6 +3,7 @@ pub mod connection;
 pub mod crypto;
 pub mod discovery;
 pub mod node;
+pub mod relay;
 pub mod stream;
 pub mod wire;
