+++ ntied/Cargo.toml
@@ -42,7 +42,6 @@ usvg = "0.36"
 tiny-skia = "0.11"
 
 [dev-dependencies]
-ntied-server = { workspace = true }
 tracing-test = "0.2.5"
 tempfile = "3.8"
 criterion = "0.5"
